% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model copy.R
\name{DcmpSepFit}
\alias{DcmpSepFit}
\title{DcmpSepFit: Decompose time series and fit models to trend and seasonality}
\usage{
DcmpSepFit(formula, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{...}{Additional arguments passed to the model fitting process.}
}
\value{
A model object of class "dcmp_sep_fit".
}
\description{
This function decomposes a time series into trend and seasonal components using X13,
and fits an ARIMA model to the trend component and a SNAIVE model to the seasonal component.
}
