% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\docType{package}
\name{fable.dcmpsepfit-package}
\alias{fable.dcmpsepfit}
\alias{fable.dcmpsepfit-package}
\title{fable.dcmpsepfit: Decompose Time Series and Fit Trend and Seasonality Models}
\description{
Allows TimeGPT forecasts from the Nixtla API to be used in a tidy workflow with the modelling interface of 'fabletools'. This extends TimeGPT to provide enhanced model specification and management, performance evaluation methods, and model combination tools.
}
\author{
\strong{Maintainer}: Your Name Your Last Name \email{your.email@example.com}

}
\keyword{package}
