% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fitted.fbl_dcmp_sep_fit}
\alias{fitted.fbl_dcmp_sep_fit}
\title{Obtain fitted values from a DcmpSepFit model}
\usage{
\method{fitted}{fbl_dcmp_sep_fit}(object, ...)
}
\arguments{
\item{object}{A DcmpSepFit model object.}

\item{...}{Additional arguments passed to the fitted method.}
}
\value{
A vector of fitted values.
}
\description{
Obtain fitted values from a DcmpSepFit model
}
