% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fitted.sep_fit}
\alias{fitted.sep_fit}
\title{Obtain fitted values from a sepfit model}
\usage{
\method{fitted}{sep_fit}(object, ...)
}
\arguments{
\item{object}{A sepfit model object.}

\item{...}{Additional arguments passed to the fitted method.}
}
\value{
A vector of fitted values.
}
\description{
Obtain fitted values from a sepfit model
}
