% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast.fbl_dcmp_sep_fit}
\alias{forecast.fbl_dcmp_sep_fit}
\title{Generate forecasts from a DcmpSepFit model}
\usage{
\method{forecast}{fbl_dcmp_sep_fit}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{A DcmpSepFit model object.}

\item{new_data}{New data to generate forecasts for.}

\item{specials}{Not used, included for compatibility.}

\item{...}{Additional arguments passed to the forecasting method.}
}
\value{
A forecasted time series object.
}
\description{
Generate forecasts from a DcmpSepFit model
}
