% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast.sep_fit}
\alias{forecast.sep_fit}
\title{SepFit模型的预测方法}
\usage{
\method{forecast}{sep_fit}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{"sep_fit"类, 是调用\code{SepFit()}得到的结果}

\item{new_data}{tsibble格式的未来数据}

\item{specials}{暂时用不到}

\item{...}{Additional arguments passed to the forecasting process}
}
\value{
A fable containing the forecasts
}
\description{
这一预测方法基于传递的trend_model、seasonal_model得到预测
因为X-13ARIMA-SEATS方法是乘法分解，所以这一方法相乘得到序列整体预测结果
}
