% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{SepFit}
\alias{SepFit}
\title{sepfit: Decompose time series and fit models to trend and seasonality}
\usage{
SepFit(formula, ...)
}
\arguments{
\item{formula}{A symbolic description of the model to be fitted.}

\item{...}{Additional arguments passed to the model fitting process.}
}
\value{
A model object of class "sep_fit".
}
\description{
This function decomposes a time series into trend and seasonal components using X13,
and fits an ARIMA model to the trend component and a SNAIVE model to the seasonal component.
}
