% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{forecast.sep_fit}
\alias{forecast.sep_fit}
\title{Forecast method for SepFit models}
\usage{
\method{forecast}{sep_fit}(object, new_data, specials = NULL, ...)
}
\arguments{
\item{object}{An object of class "sep_fit", typically the result of calling \code{SepFit()}}

\item{new_data}{A tsibble of future time points to forecast}

\item{specials}{Special arguments passed to underlying models (unused in this implementation)}

\item{...}{Additional arguments passed to the forecasting process}
}
\value{
A fable containing the forecasts
}
\description{
This function generates forecasts for SepFit models by separately forecasting
the trend and seasonal components and then combining them.
}
