% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{SepFit}
\alias{SepFit}
\title{SepFit: 分解时间序列后对子序列分别拟合}
\usage{
SepFit(formula, ...)
}
\arguments{
\item{formula}{暂时用不到}

\item{...}{暂时不支持}
}
\value{
模型类"sep_fit".
}
\description{
这一函数使用X-13ARIMA-SEATS方法分解时间序列为趋势、季节性、残差子序列
使用ARIMA和SNAIVE等方式拟合子序列并组合预测结果
}
